Fix full class for now including .cpp
